pyproject.toml
[tool.poetry]
name = "kosmos"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.30"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pytest = "^8.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
